requirements.txt
appdirs==1.4.4
black==21.6b0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.9.1
python-dotenv==0.18.0
regex==2021.7.1
toml==0.10.2
Werkzeug==2.0.1
